rename IStructure to AStructure and split structures.py
